% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filtering.R
\name{kalman_filtering}
\alias{kalman_filtering}
\title{Kalman Filtering}
\usage{
kalman_filtering(X, y, theta1, P1, Q = 0, sig = 1)
}
\arguments{
\item{X}{the explanatory variables}

\item{y}{the time series}

\item{theta1}{initial \code{theta}}

\item{P1}{initial \code{P}}

\item{Q}{(optional, default \code{0}) covariance matrix of the state noise}

\item{sig}{(optional, default \code{1}) variance of the spate noise}
}
\value{
a list containing \code{theta_arr} and \code{P_arr}, the filtered estimation of
the parameters \code{theta} and \code{P}.
}
\description{
Compute the filtered estimation of the parameters \code{theta} and \code{P}.
}
