% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood}
\usage{
loglik(X, y, Qstar, use, p1, train_theta1, train_Q, mode = "gaussian")
}
\arguments{
\item{X}{explanatory variables}

\item{y}{time series}

\item{Qstar}{the ratio \code{Q/sig^2}}

\item{use}{the availability variable}

\item{p1}{coefficient for \code{P1/sig^2 = p1 I}}

\item{train_theta1}{training set for estimation of \code{theta1}}

\item{train_Q}{time steps on which the log-likelihood is computed}

\item{mode}{(optional, default \code{gaussian})}
}
\value{
a numeric value for the log-likelihood.
}
\description{
\code{loglik} computes the log-likelihood of a state-space model of specified
\code{Q/sig^2, P1/sig^2, theta1}.
}
